Add RequestReducer, reducting transferred data by ~60%
