+++ src/Controller/DiagramController/FeatureHistoryLineChartController.ts
@@ -7,6 +7,7 @@ import DataProvider from "../Frost/DataProvider";
 import { Observation } from "../../Model/Observation";
 import { MicOutlined } from "@material-ui/icons";
 import { start } from "repl";
+import RequestReducer from "./RequestReducer";
 
 let languageProvider = Language.getInstance();
 
@@ -188,9 +189,17 @@ export class FeatureHistoryLineChartController implements IDiagramController {
         const MONTH_IN_MILLI = 1000 * 60 * 60 * 24 * 30;
         const YEAR_IN_MILLI = MONTH_IN_MILLI * 12;
         if (length >= YEAR_IN_MILLI) {
-            observations = await this.getDataByYear(end);
+            observations = await RequestReducer.getDataByYear(
+                end,
+                this.observationStation,
+                this.feature
+            );
         } else if (length >= MONTH_IN_MILLI) {
-            observations = await this.getDataByMonth(end);
+            observations = await RequestReducer.getDataByMonth(
+                end,
+                this.observationStation,
+                this.feature
+            );
         } else {
             observations = await DataProvider.getObservations(
                 this.observationStation,
@@ -239,118 +248,5 @@ export class FeatureHistoryLineChartController implements IDiagramController {
         throw new Error(
             `${name} is an invalid configuration option for a diagram of type FeatureHistoryLineChart`
         );
-    }
-
-    private async GetDataForDay(day: Date): Promise<Observation[]> {
-        const hourMilliSec = 1000 * 60 * 60;
-        var promises: Promise<Observation[]>[] = [];
-
-        day.setMinutes(0);
-
-        var zero = new Date(day.getTime());
-        var eight = new Date(day.getTime());
-        var sixteen = new Date(day.getTime());
-
-        zero.setHours(0, 0, 0);
-        eight.setHours(8, 0, 0);
-        sixteen.setHours(16, 0, 0);
-
-        promises.push(
-            DataProvider.getObservations(
-                this.observationStation,
-                this.feature,
-                zero,
-                new Date(zero.getTime() + hourMilliSec)
-            )
-        );
-        promises.push(
-            DataProvider.getObservations(
-                this.observationStation,
-                this.feature,
-                eight,
-                new Date(eight.getTime() + hourMilliSec)
-            )
-        );
-        promises.push(
-            DataProvider.getObservations(
-                this.observationStation,
-                this.feature,
-                sixteen,
-                new Date(sixteen.getTime() + hourMilliSec)
-            )
-        );
-        var r = await Promise.all(promises);
-        var results: Observation[] = [];
-        r.map((hourBlock) => {
-            if (hourBlock.length === 0) {
-                return;
-            }
-            var avg =
-                hourBlock.map((a) => a.getValue()).reduce((a, b) => a + b, 0) /
-                hourBlock.length; //avg value
-            results.push(
-                new Observation(
-                    this.observationStation,
-                    this.feature,
-                    avg,
-                    hourBlock[0].getTimeStamp()
-                )
-            );
-        });
-
-        return results;
-    }
-
-    private async getDataByMonth(endDate: Date) {
-        var startDate = new Date(endDate.getTime());
-        startDate.setMonth(startDate.getMonth() - 1);
-
-        var promises: Promise<Observation[]>[] = [];
-
-        while (startDate < endDate) {
-            promises.push(this.GetDataForDay(startDate));
-            startDate.setDate(startDate.getDate() + 1);
-        }
-        var results = await Promise.all(promises);
-        return results.reduce((prev, next) => {
-            return prev.concat(next);
-        }, []);
-    }
-
-    private async getDataByYear(endDate: Date) {
-        var observations: Promise<Observation[]>[] = [];
-        //Months
-        for (var i = 0; i < 12; i++) {
-            endDate.setDate(1);
-            observations.push(this.GetDataForDay(endDate));
-            endDate.setDate(8);
-            observations.push(this.GetDataForDay(endDate));
-            endDate.setDate(15);
-            observations.push(this.GetDataForDay(endDate));
-            endDate.setDate(22);
-            observations.push(this.GetDataForDay(endDate));
-
-            endDate.setMonth(endDate.getMonth() - 1);
-        }
-        var res = await Promise.all(observations);
-        return res.reduce((prev, next) => {
-            return prev.concat(next);
-        }, []);
-    }
-
-    private async estimateObservationCount(
-        start: Date,
-        end: Date
-    ): Promise<number> {
-        var lastHour = await DataProvider.getObservations(
-            this.observationStation,
-            this.feature,
-            new Date(Date.now() - 60 * 60 * 1000),
-            new Date(Date.now())
-        );
-        var estimate =
-            lastHour.length *
-            (((end.getTime() - start.getTime()) / 1000) * 60 * 60);
-        return estimate;
     }
 }
